requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
flask-cors==4.0.0
gunicorn==21.2.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
numpy==1.26.2
google-generativeai==0.3.1
